backend/pyproject.toml
[project]
name = "apple-health-data-backend"
version = "0.1.0"
description = "Apple Health Data Visualization - FastAPI Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "pydantic-settings>=2.6,<3",
    "webauthn>=2.1,<3",
    "python-multipart>=0.0.12,<1",
    "lxml>=5.3,<6",
    "email-validator>=2.0,<3",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28,<1",
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"
